only actually clean state files if there are candidates for removal

+++ cron/gfact_log_clean.sh
@@ -15,6 +15,6 @@ num_desc=$(ls ${state_dir}/description.* | wc -l)
 to_remove=$(find $state_dir -maxdepth 1 -name description.* -mtime +${age} | wc -l)
 
 # make sure we keep at least one reconfig around
-if [ $to_remove -lt $num_desc ];then
+if [ $to_remove -gt 0 -a $to_remove -lt $num_desc ];then
   find $state_dir -type f -mtime +${state_age} | xargs rm -f
 fi
